requirements.txt
python-telegram-bot==21.6
playwright==1.44.0
anthropic==0.30.1
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
pillow==10.4.0
httpx==0.27.0
